[DOP-8140] Split FileDownloader.Options to separated submodule
