+++ docs/file/file_downloader/options.rst
@@ -1,0 +1,11 @@
+.. _file-downloader-options:
+
+File Downloader Options
+=======================
+
+.. currentmodule:: onetl.file.file_downloader.options
+
+.. autopydantic_model:: FileDownloaderOptions
+    :member-order: bysource
+    :model-show-field-summary: false
+    :field-show-constraints: false
