+++ docs/file/file_mover/file_mover.rst
@@ -1,7 +1,7 @@
 .. _file-mover:
 
 File Mover
-==============
+==========
 
 .. currentmodule:: onetl.file.file_mover.file_mover
 
@@ -10,14 +10,7 @@ File Mover
     FileMover
     FileMover.run
     FileMover.view_files
-    FileMover.Options
 
 .. autoclass:: FileMover
     :members: run, view_files
-    :member-order: bysource
-
-.. currentmodule:: onetl.file.file_mover.file_mover.FileMover
-
-.. autopydantic_model:: Options
-    :members: mode, workers
     :member-order: bysource
