+++ docs/file/file_mover/options.rst
@@ -1,0 +1,11 @@
+.. _file-mover-options:
+
+File Mover Options
+==================
+
+.. currentmodule:: onetl.file.file_mover.options
+
+.. autopydantic_model:: FileMoverOptions
+    :member-order: bysource
+    :model-show-field-summary: false
+    :field-show-constraints: false
