+++ docs/file/file_mover/result.rst
@@ -1,9 +1,9 @@
-.. _move-result:
+.. _file-mover-result:
 
-Move result
-==============
+File Mover Result
+=================
 
-.. currentmodule:: onetl.file.file_mover.move_result
+.. currentmodule:: onetl.file.file_mover.result
 
 .. autoclass:: MoveResult
     :members: successful, failed, skipped, missing, successful_count, failed_count, skipped_count, missing_count, total_count, successful_size, failed_size, skipped_size, total_size, raise_if_failed, reraise_failed, raise_if_missing, raise_if_skipped, raise_if_empty, is_empty, raise_if_contains_zero_size, details, summary, dict, json
