+++ docs/file/file_uploader/file_uploader.rst
@@ -1,7 +1,7 @@
 .. _file-uploader:
 
 File Uploader
-==============
+=============
 
 .. currentmodule:: onetl.file.file_uploader.file_uploader
 
@@ -10,14 +10,7 @@ File Uploader
     FileUploader
     FileUploader.run
     FileUploader.view_files
-    FileUploader.Options
 
 .. autoclass:: FileUploader
     :members: run, view_files
-    :member-order: bysource
-
-.. currentmodule:: onetl.file.file_uploader.file_uploader.FileUploader
-
-.. autopydantic_model:: Options
-    :members: mode, delete_local, workers
     :member-order: bysource
