+++ docs/file/file_uploader/options.rst
@@ -1,0 +1,11 @@
+.. _file-uploader-options:
+
+File Uploader Options
+=====================
+
+.. currentmodule:: onetl.file.file_uploader.options
+
+.. autopydantic_model:: FileUploaderOptions
+    :member-order: bysource
+    :model-show-field-summary: false
+    :field-show-constraints: false
